src/core/config_loader.py
import yaml
import os

def load_config(config_path="config/coefficients.yaml"):
    """Loads the configuration from a YAML file."""
    if not os.path.exists(config_path):
        raise FileNotFoundError(f"Config file not found at {config_path}")
    
    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    return config
