pyproject.toml
[project]
name = "xp-fee-calculator"
version = "0.1.0"
description = "New Product Listing Fee Calculator"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
